+++ lib/src/effects/pulse_effect.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:skeletonizer/src/effects/painting_effect.dart';
 
 /// Creates a painting effect where a two colors are lerped
@@ -18,7 +18,7 @@ class PulseEffect extends PaintingEffect {
     super.upperBound,
     super.duration = const Duration(milliseconds: 1000),
   }) : super(reverse: true);
-
+  
   @override
   Paint createPaint(double t, Rect rect, TextDirection? textDirection) {
     final color = Color.lerp(from, to, t)!;